+++ src/peer_manager.rs
@@ -307,6 +307,7 @@ impl PeerManager {
                 evpn_instances: Vec::new(),
                 ethernet_segments: Vec::new(),
                 evpn_ip_vrfs: Vec::new(),
+                fib_tables: Vec::new(),
                 apply_bum_enforcement: false,
             },
         )
@@ -2251,6 +2252,7 @@ mod tests {
             evpn_instances: Vec::new(),
             ethernet_segments: Vec::new(),
             evpn_ip_vrfs: Vec::new(),
+            fib_tables: Vec::new(),
             apply_bum_enforcement: false,
         }
     }